from globalProvider to EthersJsProvider
