+++ common/v2/services/EthService/network/index.ts
@@ -1,3 +1,2 @@
-export { default as callProviderMethod } from './globalProvider';
-export { default as ProviderHandler } from './providerHandler';
+export { ProviderHandler } from './providerHandler';
 export { ENSProvider } from './ensProvider';
